removed save argument in pie and taylor plot
