+++ easy_mpl/__init__.py
@@ -1,15 +1,15 @@
 
 from ._main import plot
-from ._main import pie
+from ._pie import pie
 from ._main import bar_chart
-from ._main import regplot
-from ._main import imshow
+from ._regplot import regplot
+from ._imshow import imshow
 from ._main import hist
 from ._main import scatter
 from ._main import contour
 from ._main import dumbbell_plot
-from ._main import ridge
-from ._main import parallel_coordinates
+from ._ridge import ridge
+from ._parallel_coords import parallel_coordinates
 from ._taylor_plot import taylor_plot
 from ._lollipop import lollipop_plot
 from ._circular_bar_plot import circular_bar_plot
