+++ tests/test_main.py
@@ -12,11 +12,10 @@ import numpy as np
 import matplotlib.pyplot as plt
 import pandas as pd
 
-from easy_mpl import bar_chart, imshow, hist, pie, plot
-from easy_mpl import regplot, scatter, contour
+from easy_mpl import bar_chart, hist, plot
+from easy_mpl import scatter, contour
 from easy_mpl.utils import BAR_CMAPS, make_cols_from_cmap
-from easy_mpl import dumbbell_plot, ridge
-from easy_mpl import parallel_coordinates
+from easy_mpl import dumbbell_plot
 
 
 def get_chart_data(n):
@@ -116,52 +115,6 @@ class TestBarChart(unittest.TestCase):
         ax = bar_chart(np.random.randint(1, 10, 10),
                        cmap="GnBu",
                   show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-
-class TestRegplot(unittest.TestCase):
-    show = False
-    x = np.random.random(100)
-    y = np.random.random(100)
-
-    def test_reg_plot_with_line(self):
-        regplot(self.x, self.y, ci=None, show=self.show)
-        return
-
-    def test_figsize(self):
-        regplot(self.x, self.y, ci=None,
-        figsize=(10, 10),
-        show=self.show)
-        return
-
-    def test_regplot_with_line_and_ci(self):
-        regplot(self.x, self.y, show=False)
-        return
-
-    def test_regplot_with_line_ci_and_annotation(self):
-        regplot(self.x, self.y, annotation_key="MSE", annotation_val=0.2,
-                show=self.show)
-        return
-
-    def test_with_list_as_inputs(self):
-        regplot(self.x.tolist(), self.y.tolist(),
-                show=self.show)
-        return
-
-    def test_with_nan_vals(self):
-        x = np.random.random(100)
-        x[10] = np.nan
-        y = np.random.random(100)
-        y[5] = np.nan
-        ax = regplot(x, y, show=self.show, title='test_with_nan_vals')
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nan_in_x(self):
-        x = np.append(self.x, np.nan)
-        y = np.append(self.y, 0.5)
-        ax = regplot(x, y, show=self.show, title='test_with_nan_vals')
         assert isinstance(ax, plt.Axes)
         return
 
@@ -280,83 +233,6 @@ class TestPlot(unittest.TestCase):
         return
 
 
-class TestImshow(unittest.TestCase):
-    show = False
-
-    def test_imshow(self):
-        ax, _ = imshow(np.random.random((10, 10)), colorbar=True, title="vanilla",
-                       show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_figsize(self):
-        ax, _ = imshow(np.random.random((10, 10)), colorbar=True,
-            title="figsize", figsize=(10, 10),
-                       show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_imshow_witout_cb(self):
-        ax, img= imshow(np.random.random((10, 10)), colorbar=False, title="without_colorbar",
-                        show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_annotate(self):
-        ax, img= imshow(np.random.random((10, 10)), colorbar=False, title="annotate",
-                        annotate=True, show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_df(self):
-        df = pd.DataFrame(np.random.random((10, 2)), columns=['a', 'b'])
-        ax, img = imshow(df, colorbar=True, show=self.show, title="df")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_with_nan_vals(self):
-        x = np.random.random((10, 10))
-        x.ravel()[np.random.choice(x.size, 5, replace=False)] = np.nan
-        ax, img = imshow(x, colorbar=True, show=self.show, title="with_nan_vals")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_white_gridlines(self):
-        data = np.random.random((4, 10))
-        ax, im = imshow(data, cmap="YlGn",
-                        xticklabels=[f"Feature {i}" for i in range(data.shape[1])],
-                        white_grid=True, annotate=True, colorbar=True, show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_df_object_type(self):
-        data = pd.DataFrame(np.random.random((10, 10)), dtype='object')
-        ax, im = imshow(data, white_grid=True, annotate=True, show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_textcolors_tuple(self):
-        data = pd.DataFrame(np.random.random((10, 10)), dtype='object')
-        ax, im = imshow(data, white_grid=True, annotate=True, show=self.show,
-                        annotate_kws={'textcolors': ("black", "white")}
-                        )
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_textcolors_array(self):
-        data = pd.DataFrame(np.random.random((3, 3)), dtype='object')
-        ax, im = imshow(data, cmap="YlGn",
-       white_grid=True, annotate=True,
-       annotate_kws={
-              "textcolors": np.array([['black', 'black', 'black'],
-                                      ['white', 'white', 'white'],
-                                     ['green', 'green', 'green']])
-       },
-       show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-
 class Testhist(unittest.TestCase):
     show = False
 
@@ -379,44 +255,6 @@ class Testhist(unittest.TestCase):
         x = np.random.random((10, 1))
         x.ravel()[np.random.choice(x.size, 5, replace=False)] = np.nan
         ax = hist(x, show=self.show, title="with_nan_vals")
-        assert isinstance(ax, plt.Axes)
-        return
-
-
-class TestPie(unittest.TestCase):
-    show = False
-    save = False
-
-    def test_binary(self):
-        ax = pie(np.random.randint(0, 2, 100), show=self.show, save=self.save)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_multiclass(self):
-        ax = pie(np.random.randint(0, 5, 100), show=self.show, save=self.save)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_string(self):
-        ax = pie(['a'] * 60 + ['b'] * 50, show=self.show, save=self.save)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_fraction(self):
-        ax = pie([0.1, 0.2, 0.5, 0.2], show=self.show, save=self.save)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nan_in_fraction(self):
-        ax = pie([0.1, 0.2, np.nan, 0.2], show=self.show, save=self.save,
-                 title="nan_in_fraction")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nan_in_vals(self):
-        vals = np.random.randint(0, 5, 100).astype("float32")
-        vals[np.random.choice(vals.size, 5, replace=False)] = np.nan
-        ax = pie(vals, show=self.show, save=self.save, title="nan_in_vals")
         assert isinstance(ax, plt.Axes)
         return
 
@@ -577,203 +415,6 @@ class TestDumbbell(unittest.TestCase):
         en[4] = np.nan
         ax = dumbbell_plot(st, en, show=self.show,
                             title="with labels")
-        assert isinstance(ax, plt.Axes)
-        return
-
-
-class TestRidge(unittest.TestCase):
-    show = False
-
-    def test_df_3_cols(self):
-        df = pd.DataFrame(np.random.random((100, 3)), dtype='object')
-        axis = ridge(df, title="df_3_cols", show=self.show)
-        for ax in axis:
-            assert isinstance(ax, plt.Axes)
-        return
-
-    def test_df_20_cols(self):
-        df = pd.DataFrame(np.random.random((100, 20)))
-        axis = ridge(df, title="df_20_cols", show=self.show)
-        for ax in axis:
-            assert isinstance(ax, plt.Axes)
-        return
-
-    def test_df_1col(self):
-        df = pd.DataFrame(np.random.random((100, 1)))
-        axis = ridge(df, title="df_1col", show=self.show)
-        for ax in axis:
-            assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nparray_2d(self):
-        axis = ridge(np.random.random((100, 5)), title="nparray_2d", show=self.show)
-        for ax in axis:
-            assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nparray_1d(self):
-        axis = ridge(np.random.random(100,), title="nparray_1d", show=self.show)
-        for ax in axis:
-            assert isinstance(ax, plt.Axes)
-        return
-
-    def test_array_with_nan_vals(self):
-        x = np.random.random(100,).astype("float32")
-        x[np.random.randint(0, 100, 10)] = np.nan
-        axis = ridge(x, title="array_with_nan_vals", show=self.show)
-        for ax in axis:
-            assert isinstance(ax, plt.Axes)
-        return
-
-
-class TestParallelCoord(unittest.TestCase):
-    show = False
-    ynames = ['P1', 'P2', 'P3', 'P4', 'P5']
-
-    N1, N2, N3 = 10, 5, 8
-    N = N1 + N2 + N3
-
-    y1 = np.random.uniform(0, 10, N) + 7
-    y2 = np.sin(np.random.uniform(0, np.pi, N))
-    y3 = np.random.binomial(300, 1 / 10, N)
-    y4 = np.random.binomial(200, 1 / 3, N)
-    y5 = np.random.uniform(0, 800, N)
-
-    data = np.column_stack((y1, y2, y3, y4, y5))
-    data = pd.DataFrame(data, columns=ynames)
-
-    def test_catfeatures(self):
-        data = self.data.copy()
-        data['P5'] = random.choices(['a', 'b', 'c', 'd', 'e', 'f'], k=self.N)
-        ax = parallel_coordinates(data, names=self.ynames, title="cat feature",
-                             show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_customticklabelscatfeatures(self):
-        data = self.data.copy()
-        data['P5'] = random.choices(['a', 'b', 'c', 'd', 'e', 'f'], k=self.N)
-        ax = parallel_coordinates(data, title="custom ticklabels cat feat",
-                                  ticklabel_kws={"fontsize": 8, "color": "red"},
-                                  show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_allcontinuous(self):
-        data = self.data.copy()
-        ax = parallel_coordinates(data, names=self.ynames, title="all continuous",
-                             show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_continuoustarget(self):
-        data = self.data.copy()
-        ax = parallel_coordinates(data, names=self.ynames,
-                                  categories=np.random.randint(0, 5, self.N),
-                                  title="target continuous",
-                                  show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_targetcategory(self):
-        category = ['a', 'b', 'c', 'd', 'e']
-        data = self.data.copy()
-        ax = parallel_coordinates(data, names=self.ynames,
-                                  categories=random.choices(category, k=len(data)),
-                                  show=self.show,
-                         title="target category")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nparray(self):
-        data = self.data.copy()
-        ax = parallel_coordinates(data.values, names=self.ynames,
-                             title="np array",
-                             show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_customticklabels(self):
-        data = self.data.copy()
-        ax = parallel_coordinates(data.values,
-                             title="custom ticklabels",
-                             ticklabel_kws={"fontsize": 8, "color": "red"},
-                             show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_straightlines(self):
-        data = self.data.copy()
-        ax = parallel_coordinates(data, linestyle="straight",
-                             title="straight lines",
-                             show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_all_cat(self):
-        data = {
-            'tide': ['yeo', 'scale', 'log', 'minmax', 'robust'],
-            'wat': ['scale', 'log', 'sqrt', 'quantile', 'log'],
-            'estimator': ['log', 'zscore', 'log', 'robust', 'zscore'],
-            'P4': ['lr', 'lasso', 'lr', 'lr', 'lasso'],
-        }
-        cat = [3.10e14, 1.15e14, 1.20e14, 1.25e14, 1.50e14]
-
-        data = pd.DataFrame.from_dict(data)
-
-        ax = parallel_coordinates(data, cat, show=self.show, title="all cat")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_all_cat_but_one(self):
-        data = {
-            'tide': ['yeo', 'scale', 'log', 'minmax', 'robust'],
-            'wat': ['scale', 'log', 'sqrt', 'quantile', 'log'],
-            'estimator': ['log', 'zscore', 'log', 'robust', 'zscore'],
-            'P3': [11,2,33,4, 5],
-            'P4': ['lr', 'lasso', 'lr', 'lr', 'lasso'],
-        }
-        cat = [3.10e14, 1.15e14, 1.20e14, 1.25e14, 1.50e14]
-
-        data = pd.DataFrame.from_dict(data)
-
-        ax = parallel_coordinates(data, cat, show=self.show, title="all cat but one")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nan_in_data(self):
-        data = self.data.copy()
-        data.iloc[4, 4] = np.nan
-        data.iloc[3, 3] = np.nan
-        ax = parallel_coordinates(data, names=self.ynames,
-                                  title="nan in data",
-                                  show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nan_in_categories(self):
-
-        data = self.data.copy()
-        categories = np.arange(len(data)).astype("float32")
-        categories[4] = np.nan
-
-        ax = parallel_coordinates(data, names=self.ynames,
-                                  categories=categories,
-                                  title="nan in categories",
-                                  show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_nan_in_data_and_categroes(self):
-        data = self.data.copy()
-        data.iloc[4, 4] = np.nan
-        data.iloc[3, 3] = np.nan
-        categories = np.arange(len(data)).astype("float32")
-        categories[6] = np.nan
-        ax = parallel_coordinates(data, names=self.ynames,
-                                  categories=categories,
-                                  title="nan in data and categories",
-                                  show=self.show)
         assert isinstance(ax, plt.Axes)
         return
